requirements.txt
streamlit
google-genai
Authlib>=1.3.2
google-cloud-storage